Modifications for ERScript_r2 release
ERScript_r2 will include ERmodel_v1.4 which is characterised by a version of the meta-model in which attributes are modelled as attributes and there are no enumeration_types for attributes. The transforms and code generators work via an intermediate step back into the v1.3 format. This release is now debugged and working.